Functions Import
Add
Functions import
Fix miniBugs
Little Clean Code
